add DEPLOYMENT env var, make migrations idempotent in prod

+++ db/migrate/20150618142658_create_salesforce_schema.rb
@@ -1,9 +1,11 @@
 class CreateSalesforceSchema < ActiveRecord::Migration
   def up
+    return if ENV["DEPLOYMENT"] != "development"
     execute "CREATE SCHEMA salesforce"
   end
 
   def down
+    return if ENV["DEPLOYMENT"] != "development"
     execute "DROP SCHEMA salesforce CASCADE"
   end
 end
